backend/requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
cssselect2==0.7.0
defusedxml==0.8.0rc2
Django==4.2.6
django-extra-fields==3.0.2
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-extra-fields==3.7.0
filetype==1.2.0
fonttools==4.43.1
gobject==0.1.0
gunicorn==21.2.0
html5lib==1.1
idna==3.4
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pybase64==1.3.1
pycparser==2.21
psycopg2_binary==2.9.9
pydyf==0.8.0
PyJWT==2.8.0
pyphen==0.14.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
reportlab==4.0.6
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
tinycss2==1.2.1
urllib3==2.0.7
weasyprint==60.1
webencodings==0.5.1
zopfli==0.2.3
